+++ gems/aws-sdk-workspaces/CHANGELOG.md
@@ -1,6 +1,11 @@
 Unreleased Changes
 ------------------
 
+1.25.0 (2019-06-28)
+------------------
+
+* Feature - Minor API fixes for WorkSpaces.
+
 1.24.0 (2019-06-27)
 ------------------